PCAD/pyproject.toml
[project]
name = "pcad"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "jupyterlab>=4.5.6",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "mysql-connector-python>=9.6.0",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pymongo>=4.16.0",
    "scikit-learn>=1.8.0",
    "scipy-stubs~=1.17.1",
    "seaborn>=0.13.2",
    "seaborn-image>=0.10.0",
]
